% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntTrap.R
\name{IntTrap}
\alias{IntTrap}
\title{Integral por la regla del trapecio.}
\usage{
IntTrap(f, a, b, n = 1)
}
\arguments{
\item{f}{Función a integrar.}

\item{a}{Origen del intervalo de integración.}

\item{b}{Final el intervalo de integración.}

\item{n}{Número de subintervalos donde aplicar el método (por defecto n=1).}
}
\value{
Devuelve la integral de f entre a y b realizada por la regla del trapecio sobre n subintervalos equidistantes entre a y b.
}
\description{
Integral por la regla del trapecio.
}
\examples{
IntTrap(cos,0,1), IntTrap(cos,0,1,n=4)

}
\seealso{
\link{IntRect}, \link{IntPM}, \link{IntSimpson}
}
